+++ source/Databricks/documents/release-notes/release-notes.md
@@ -1,5 +1,10 @@
 # Databricks Release Notes
 
+## Version 7.2.0
+
+- Databricks streaming of `ApacheArrow` and `JsonArray` is added.
+- The `DatabricksSqlStatementClient` is deprecated. Please use `DatabricksSqlWarehouseQueryExecutor` instead. See [documentation.md](../documentation.md) for information on how to use the new functionality.
+
 ## Version 7.1.2
 
 - The "Internal" namespace is removed, and related code is grouped in relevant namespaces.